format

+++ app/src/main/java/com/github/dawidhyzy/kotlinvsjava/jv/api/model/Rain.java
@@ -5,13 +5,13 @@ import com.google.gson.annotations.SerializedName;
 public class Rain {
 
     @SerializedName("3h")
-    private int _3h;
+    private double _3h;
 
-    public int get3h() {
+    public double get3h() {
         return _3h;
     }
 
-    public void set3h(int _3h) {
+    public void set3h(double _3h) {
         this._3h = _3h;
     }
